requirements.txt
alembic==1.11.1
anyio==3.7.0
apache-libcloud==3.7.0
async-timeout==4.0.2
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.0
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.95.2
fastapi-users==11.0.0
fastapi-users-db-sqlalchemy==5.0.0
fasteners==0.18
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.2
orjson==3.8.14
passlib==1.7.4
pycparser==2.21
pydantic==1.10.8
PyJWT==2.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
redis==4.5.5
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.15
sqlalchemy-file==0.4.0
starlette==0.27.0
typing_extensions==4.6.2
ujson==5.7.0
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
